Update Compunded Interest Python.py

+++ 2019/Compunded Interest Python.py
@@ -4,14 +4,14 @@ r= input('Enter value of Interest Rate')
 
 p= int(p)
 n= int(n)
-r= int(r)
+r= float(r)
 
 r= (r/100)+1
 
 ci= p*((r**n)-1)
 
 print ("You will earn:",ci, 'in return')
-ci= int(ci)
+ci= float(ci)
 cit= p+ci
 print ("You will then have a total of:",cit, 'available')
 input ('Press ENTER to exit')
